+++ app/types/unit.ts
@@ -1,0 +1,11 @@
+export type Unit =
+	| "GRAM"
+	| "KILOGRAM"
+	| "MILLILITER"
+	| "CENTILITER"
+	| "LITER"
+	| "PIECE"
+	| "TABLESPOON"
+	| "TEASPOON"
+	| "CUP";
+
